% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify_data}
\alias{deidentify_data}
\title{De-identify data through encrypting text columns, grouping rare values, and,
aggregating numeric or date columns.}
\usage{
deidentify_data(
  data,
  cols_to_encrpyt = NULL,
  seeds_for_encryption = NULL,
  group_rare_values_cols = NULL,
  group_rare_values_limit = NULL,
  group_rare_values_text = NA,
  time_aggregation = c("week", "month", "bimonth", "quarter", "halfyear", "year"),
  quiet = FALSE
)
}
\arguments{
\item{data}{A data.frame with the data you want to deidentify.}

\item{cols_to_encrpyt}{A string or vector of strings with the columns that you want to encrypt
using the \code{seed_cipher()} function from the \code{caesar} package.}

\item{seeds_for_encryption}{A string or vector of strings with the columns that you want to use for
each column in the cols_to_encrpyt parameter. If NULL, will use random seeds
and will print those seeds if the parameter quiet is set to TRUE.}

\item{group_rare_values_cols}{A string or vector of strings with the columns that you want to convert
rare values (below a certain percent of all values as set in
\code{group_rare_values_limit}) into NA or a particular string (or NA) set in
\code{group_rare_values_text}.}

\item{group_rare_values_limit}{A string or vector of strings (one for each col in \code{group_rare_values_cols})
for what threshold (in percent of all non-NA values) to determine that a value
is rare enough to change to NA (or the string set in \code{group_rare_values_text}).}

\item{group_rare_values_text}{A string or vector of strings (one for each col in \code{group_rare_values_cols})
for what to rename the values that are determined to be rare enough (based on
threshold set in \code{group_rare_values_limit} to rename (default is just
calling them NA).}

\item{time_aggregation}{A string with the time unit to aggregate all Date variables to.
Can take one of the following: 'week', 'month', 'bimonth', 'quarter',
'halfyear', 'year'.}

\item{quiet}{A Boolean for whether you want to output a message that tells you which
columns that you are encrypting and the seed set for each column to
do the encryption. If you don't set the seed yourself, you need these
seeds to decrypt.}
}
\value{
A data.frame with the selected columns de-identify based on user parameters.
}
\description{
De-identify data through encrypting text columns, grouping rare values, and,
aggregating numeric or date columns.
}
