% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deidentify_example}
\alias{deidentify_example}
\title{A random selection of 10,000 hate crimes in the United States}
\format{
A data frame with 10000 rows and 21 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\usage{
deidentify_example
}
\description{
A dataset containing details about 10,000 hate crimes in the United States.
}
\keyword{datasets}
