% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_keys.R
\name{gen_aes_key}
\alias{gen_aes_key}
\title{generate aes or libsodium keys}
\usage{
gen_aes_key(...)
}
\arguments{
\item{...}{arguments to pass to openssl::aes_keygen or sodium::keygen}
}
\value{
return a cyphr key object for use with deidentify_id and identify_id functions
}
\description{
generate aes or libsodium keys
}
\examples{
\dontrun{gen_aes_key(12)}
}
