% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_keys.R
\name{gen_sodium_key}
\alias{gen_sodium_key}
\title{Generate sodium key}
\usage{
gen_sodium_key(...)
}
\arguments{
\item{...}{arguments to pass to \code{openssl::aes_keygen()} or \code{sodium::keygen()}}
}
\value{
A \code{cyphr} key object for use with \code{deidentify_id()} and identify_id functions
}
\description{
Generate sodium key
}
\examples{
\dontrun{key_sodium(12)}
}
