% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_aggregate_dates}
\alias{graph_aggregate_dates}
\title{Create a barplot showing the number of unique dates at each level of aggregation}
\usage{
graph_aggregate_dates(data)
}
\arguments{
\item{data}{A vector of dates in the Date format.}
}
\value{
A \code{ggplot2} graph
}
\description{
This takes a vector of Dates and creates a barplot showing how many unique dates are in the data, at the original inputted unit and at every level of aggregation (week, month, bimonth, quarter, halfyear, year).
}
\examples{
graph_aggregate_dates(deidentify::initiations$received_date)
}
