% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_group_rare_values}
\alias{graph_group_rare_values}
\title{Graph how much data is dropped for each percent limit to remove rare values}
\usage{
graph_group_rare_values(data)
}
\arguments{
\item{data}{A vector of strings or numbers.}
}
\value{
A \code{ggplot2} graph
}
\description{
Makes a graph showing what percent of values are dropped when excluding values
that are less common than k\% of the data for every value of k from 1\%-100\%.
}
\examples{
graph_group_rare_values(mtcars$mpg)
}
