% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hate_crimes}
\alias{hate_crimes}
\title{A random selection of 1,000 hate crimes in the United States}
\format{
A data frame with 1000 rows and 7 variables:
\describe{
\item{ori}{A unique identifier code for the police agency}
\item{incident_date}{The date of the incident}
\item{population}{The population of the city where the police agency has jurisdiction (generally the city or county population)}
\item{agency_name}{The name of the police agency}
\item{unique_id}{A unique identifier for the incident}
\item{ucr_offense_code_1}{The crime that occurred (e.g. murder)}
\item{bias_motivation_offense_1}{The type of hate crime (e.g. anti-Black)}
...
}
}
\usage{
hate_crimes
}
\description{
A dataset containing details about 1,000 hate crimes in the United States.
}
\keyword{datasets}
