% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_methods.R
\name{identify_text}
\alias{identify_text}
\title{decrypt data using a cyphr key}
\usage{
identify_text(.data, key, cols_to_decrypt)
}
\arguments{
\item{.data}{a data frame}

\item{key}{A cyphr key generated from the cyphr package}

\item{cols_to_decrypt}{Specific to data.frames. Should be a character vector of the names of the columns that you
would like to decrypt.}
}
\value{
either a data.frame or character vector depending on the object passed through
}
\description{
When you encrypt a field with deidentify, it is relatively secure depending on the method of encryption
that you may use. However, in the case that you need to decrypt any of the fields and used encryption
methods to do so, you will also be able to decrypt any of the fields back to their original form.
}
\examples{
\dontrun{
key <- gen_aes_key()
encrypt_string <- deidentify_text("cat",key)
identical(identify_text(encrypt_string,key),"cat")
}
}
