% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initiations}
\alias{initiations}
\title{Cook County Initiations Data}
\format{
A tibble with 1000 rows and 38 variable
\describe{
for in depth definitions, look at \url{https://datacatalog.cookcountyil.gov/Courts/Initiation/7mck-ehwz}}
}
\source{
\url{https://datacatalog.cookcountyil.gov/Courts/Initiation/7mck-ehwz}
}
\usage{
initiations
}
\description{
This is a sample of Cook County's Initiations data available on the Cook County Open Data Portal
}
\keyword{datasets}
