% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_deidentify_choices_reported}
\alias{make_deidentify_choices_reported}
\title{Title}
\usage{
make_deidentify_choices_reported(
  data,
  drop_rare_values_cols,
  preferred_k_score = NULL,
  file_name
)
}
\arguments{
\item{file_name}{}
}
\value{

}
\description{
Title
}
