% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_k_score.R
\name{make_k_score}
\alias{make_k_score}
\title{Make k score out of de-identified data}
\usage{
make_k_score(data, columns = NULL, quiet = FALSE, minimum_k_score = 3)
}
\arguments{
\item{data}{The data.frame with the data you want to make the k score from.}

\item{columns}{A string or vector of strings for the names of columns to group the data by.}

\item{quiet}{A Boolean (default FALSE) for whether you want to output a message that tells you if
there are groups with too few observations - you can set how many is 'too few'
in the parameter \code{minimum_k_score}, default is 3.}

\item{minimum_k_score}{A single integer which sets the number of observations in each group that will
message (if \code{quiet} is FALSE) saying how many groups have fewer observations below
that group.}
}
\value{
A data.frame
}
\description{
This gives you a data.frame with rows grouped based on the columns
parameter and tell you how many unique values are in each group.
}
\examples{
make_k_score(mtcars, columns = c("cyl", "vs", "am", "gear"))
}
