% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{deidentify_choices_table}
\alias{deidentify_choices_table}
\title{Make a data.frame showing either the outcome (k-score) of all possible de-identify choices or only those that meet a certain k-score threshold.}
\usage{
deidentify_choices_table(
  data,
  group_rare_values_cols,
  preferred_k_score = NULL
)
}
\arguments{
\item{data}{A data.frame with the data you want to de-identify.}

\item{group_rare_values_cols}{A string or vector of strings with the columns that you want to convert
rare values (below a certain percent of all values as set in
\code{group_rare_values_limit}) into NA or a particular string (or NA) set in
\code{group_rare_values_text}.}

\item{preferred_k_score}{A number of vector of numbers to set the minimum (and maximum if a vector) k-score you want from the possible choices.}
}
\value{
Returns a data.frame that only has all possible choices of decisions to make and the k-score that it returns. Each row is a possible decision when using the \code{deidentify_data()} function and includes summary statistics for the k-score of that decision. If preferred_k_score is set, returns only choices that meet this parameter. If no choices meet this k-score minimum, will return an empty data.frame.
}
\description{
Make a data.frame showing either the outcome (k-score) of all possible de-identify choices or only those that meet a certain k-score threshold.
}
