% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify_methods.R
\name{deidentify_text}
\alias{deidentify_text}
\title{methods to deidentify different forms of data}
\usage{
deidentify_text(.data, cols_to_encrypt, key = NULL)
}
\arguments{
\item{.data}{a data frame with columns to encrypt}

\item{cols_to_encrypt}{A character vector with the names associated with the columns to encrypt}

\item{key}{A cyphr key generated from the cyphr package}
}
\value{
A Dataframe with encrypted columns ending with the prefix 'encrypt_id'
}
\description{
methods to deidentify different forms of data
}
\examples{
\dontrun{
}
}
