% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify_methods.R
\name{deidentify_text}
\alias{deidentify_text}
\title{methods to deidentify different forms of data}
\usage{
deidentify_text(.data, cols_to_encrypt, key = NULL)
}
\arguments{
\item{.data}{a data frame}

\item{cols_to_encrypt}{A character vector with the names associated with the columns to encrypt}

\item{key}{A cyphr key generated from the cyphr package}
}
\value{
A Dataframe with encrypted columns ending with the prefix 'encrypt_id'
}
\description{
The deidentify_text method offers a flexible approach to encryption offering the user the option to encrypt text
columns using either openssl or sodium. The encryption path chosen depends on the encryption key generated by
the user. For example, if a user generates an aes_key, and passes it to deidentify_text, aes encryption will be
automatically applied. The function will return a dataframe with the an encrypt_id suffix added to the end of the
column name. Thus, a user will have to actively remove the unencrypted columns manually.
}
\examples{
\dontrun{
data("initiations")
aes_key <- gen_aes_key(24)
deidentify_text(initiations,"case_participant_id",aes_key)
}
}
