% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_case_score.R
\name{make_case_score}
\alias{make_case_score}
\title{Title}
\usage{
make_case_score(data, columns, quiet = FALSE)
}
\arguments{
\item{data}{The data.frame with the data you want to make the case score from.}

\item{columns}{A string or vector of strings for the names of columns to group the data by.}

\item{quiet}{A Boolean for whether you want to output a message that tells you if
there are groups with 3 or fewer observations (default is FALSE meaning
that it is NOT quiet, and thus will give the message).}
}
\value{
A data.frame
}
\description{
Title
}
\examples{
make_case_score(mtcars, columns = c("cyl", "vs", "am", "gear"))
}
