% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_deidentify_choices_reported}
\alias{make_deidentify_choices_reported}
\title{Produce a PDF showing tables of possible de-identification choices, choices that meet a set parameter, and graphs of each decision for individual variables.}
\usage{
make_deidentify_choices_reported(
  data,
  group_rare_values_cols,
  preferred_k_score = NULL,
  file_name
)
}
\arguments{
\item{data}{A data.frame with the data you want to de-identify.}

\item{group_rare_values_cols}{A string or vector of strings with the columns that you want to convert
rare values (below a certain percent of all values as set in
\code{group_rare_values_limit}) into NA or a particular string (or NA) set in
\code{group_rare_values_text}.}

\item{preferred_k_score}{A number of vector of numbers to set the minimum (and maximum if a vector) k-score you want from the possible choices.}

\item{file_name}{A string with the name of the PDF file that will be generated by the function. Must end in '.pdf'. The file will be generated in the current working directory.}
}
\value{
A PDF with a table of all possible choices, a table of the choices that meet the \code{preferred_k_score} parameter, and graphs of decision outcomes.
}
\description{
Produce a PDF showing tables of possible de-identification choices, choices that meet a set parameter, and graphs of each decision for individual variables.
}
